+++ src/utils/vec3_rotation.c
@@ -3,39 +3,39 @@
 /*                                                        :::      ::::::::   */
 /*   vec3_rotation.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: nmihaile <nmihaile@student.42heilbronn.    +#+  +:+       +#+        */
+/*   By: bwerner <bwerner@student.42heilbronn.de>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/02 14:47:52 by nmihaile          #+#    #+#             */
-/*   Updated: 2024/08/02 22:25:18 by nmihaile         ###   ########.fr       */
+/*   Updated: 2024/08/10 20:14:58 by bwerner          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../inc/miniRT.h"
 
-t_vec3	vec3_rotate_x(t_vec3 v, double rad)
+t_vec3	vec3_rotate_x(t_vec3 v, float rad)
 {
-	double	s;
-	double	c;
+	float	s;
+	float	c;
 
 	s = sin(rad);
 	c = cos(rad);
 	return ((t_vec3){v.x, v.y * c - v.z * s, v.y * s + v.z * c});
 }
 
-t_vec3	vec3_rotate_y(t_vec3 v, double rad)
+t_vec3	vec3_rotate_y(t_vec3 v, float rad)
 {
-	double	s;
-	double	c;
+	float	s;
+	float	c;
 
 	s = sin(rad);
 	c = cos(rad);
 	return ((t_vec3){v.x * c + v.z * s, v.y, -v.x * s + v.z * c});
 }
 
-t_vec3	vec3_rotate_z(t_vec3 v, double rad)
+t_vec3	vec3_rotate_z(t_vec3 v, float rad)
 {
-	double	s;
-	double	c;
+	float	s;
+	float	c;
 
 	s = sin(rad);
 	c = cos(rad);
